feat(build): add BUILD_TESTS option and --no-tests flag to build.sh
CMakeLists.txt:
- Add BUILD_TESTS option (default ON)
- Wrap all test/benchmark targets in BUILD_TESTS guard

build.sh:
- Add 'no-tests' argument to skip test targets during refactoring
- Usage: ./build.sh Release no-tests

+++ include/mail_system/back/mailServer/session/session_base.tpp
@@ -265,7 +265,7 @@ bool SessionBase<ConnectionType>::record_auth_failure_and_check() {
         auto cfg = std::atomic_load(&m_server->m_config);
         if (cfg) max_attempts = cfg->max_auth_attempts;
     }
-    if (auth_attempt_count_ >= max_attempts) {
+    if (static_cast<size_t>(auth_attempt_count_) >= max_attempts) {
         LOG_SESSION_WARN("AUTH failures exceeded ({}/{}), closing session from {}",
             auth_attempt_count_, max_attempts, get_client_ip());
         return true;
